+++ src/Concordium/Client/Runner.hs
@@ -94,7 +94,6 @@ import Control.Monad.State.Strict
 import Data.Aeson as AE
 import qualified Data.Aeson.Encode.Pretty as AE
 import qualified Data.ByteString as BS
-import qualified Data.ByteString.Builder as BSBuilder
 import qualified Data.ByteString.Lazy as BSL
 import qualified Data.ByteString.Lazy.Char8 as BSL8
 import qualified Data.ByteString.Short as BS (toShort)
@@ -920,6 +919,10 @@ processTransactionCmd action baseCfgDir verbose backend =
                     handlePLTModifyList backend baseCfgDir verbose modifyListAction account tokenId txOpts
                 TransactionPLTPausation pauseAction tokenId txOpts ->
                     handlePLTPausation backend baseCfgDir verbose pauseAction tokenId txOpts
+                TransactionPLTModifyAdminRoles adminAction roles account tokenId txOpts ->
+                    handlePLTModifyAdminRoles backend baseCfgDir verbose adminAction roles account tokenId txOpts
+                TransactionPLTUpdateMetadata metadataURL maybeChecksum tokenId txOpts ->
+                    handlePLTUpdateMetadata backend baseCfgDir verbose metadataURL maybeChecksum tokenId txOpts
 
 -- | Renormalize a 'TokenAmount' to conform to the number of decimal places expected by the
 --  token. If more than the expected number of decimals are given, this fails with an error.
@@ -979,9 +982,9 @@ handlePLTTransfer backend baseCfgDir verbose receiver amount tokenIdText maybeMe
         let tokenTransfer = CBOR.TokenTransfer tokenTransferBody
         let tokenUpdateTransaction = CBOR.TokenUpdateTransaction (Seq.singleton tokenTransfer)
         let bytes = CBOR.tokenUpdateTransactionToBytes tokenUpdateTransaction
-        let tokenParameter = Types.TokenParameter $ BS.toShort bytes
+        let operations = Types.rawCborFromBytes bytes
 
-        let payload = Types.TokenUpdate tokenId tokenParameter
+        let payload = Types.TokenUpdate tokenId operations
         let encodedPayload = Types.encodePayload payload
 
         let nrgCost _ = return $ Just $ flip (tokenUpdateTransactionEnergyCost (Types.payloadSize encodedPayload) Cost.tokenTransferCost) $ extendedCostFromOpts txOpts
@@ -1023,9 +1026,9 @@ handlePLTUpdateSupply backend baseCfgDir verbose tokenSupplyAction amount tokenI
 
         let tokenUpdateTransaction = CBOR.TokenUpdateTransaction (Seq.singleton tokenOperation)
         let bytes = CBOR.tokenUpdateTransactionToBytes tokenUpdateTransaction
-        let tokenParameter = Types.TokenParameter $ BS.toShort bytes
+        let operations = Types.rawCborFromBytes bytes
 
-        let payload = Types.TokenUpdate tokenId tokenParameter
+        let payload = Types.TokenUpdate tokenId operations
         let encodedPayload = Types.encodePayload payload
 
         let opCost
@@ -1066,16 +1069,108 @@ handlePLTModifyList backend baseCfgDir verbose modifyListAction account tokenIdT
 
         let tokenUpdateTransaction = CBOR.TokenUpdateTransaction (Seq.singleton tokenOperation)
         let bytes = CBOR.tokenUpdateTransactionToBytes tokenUpdateTransaction
-        let tokenParameter = Types.TokenParameter $ BS.toShort bytes
+        let operations = Types.rawCborFromBytes bytes
 
         tokenId <- case tokenIdFromText tokenIdText of
             Right val -> return val
             Left err -> logFatal ["Error couldn't parse token id:", err]
 
-        let payload = Types.TokenUpdate tokenId tokenParameter
+        let payload = Types.TokenUpdate tokenId operations
         let encodedPayload = Types.encodePayload payload
 
         let nrgCost _ = return $ Just $ flip (tokenUpdateTransactionEnergyCost (Types.payloadSize encodedPayload) Cost.tokenListOperationCost) $ extendedCostFromOpts txOpts
+        txCfg <- liftIO $ getTransactionCfg baseCfg txOpts nrgCost
+
+        let intOpts = toInteractionOpts txOpts
+        let outFile = toOutFile txOpts
+        signAndProcessTransaction_ verbose txCfg encodedPayload intOpts outFile backend
+
+handlePLTModifyAdminRoles ::
+    Backend ->
+    Maybe FilePath ->
+    Bool ->
+    ModifyAdminAction ->
+    [CBOR.TokenAdminRole] ->
+    Text ->
+    Text ->
+    TransactionOpts (Maybe Types.Energy) ->
+    IO ()
+handlePLTModifyAdminRoles backend baseCfgDir verbose adminAction roles account tokenIdText txOpts = do
+    baseCfg <- getBaseConfig baseCfgDir verbose
+    when verbose $ do
+        runPrinter $ printBaseConfig baseCfg
+        putStrLn ""
+
+    adminAddress <- getAccountAddressArg (bcAccountNameMap baseCfg) account
+    let cborAdminAddress = CBOR.accountTokenHolder $ naAddr adminAddress
+
+    let updateAdminRolesDetails = CBOR.UpdateAdminRolesDetailsBuilder (Just cborAdminAddress) (Just (Seq.fromList roles))
+    let eitherUpdateAdminRolesDetails = CBOR.buildUpdateAdminRolesDetails updateAdminRolesDetails
+    updateAdminRolesDetailsBody <- case eitherUpdateAdminRolesDetails of
+        Right val -> return val
+        Left err -> logFatal ["Error creating token update admin roles details:", err]
+
+    withClient backend $ do
+        tokenOperation <- case adminAction of
+            AssignAdminRole -> pure $ CBOR.TokenAssignAdminRoles updateAdminRolesDetailsBody
+            RevokeAdminRole -> pure $ CBOR.TokenRevokeAdminRoles updateAdminRolesDetailsBody
+        let tokenUpdateTransaction = CBOR.TokenUpdateTransaction (Seq.singleton tokenOperation)
+        let bytes = CBOR.tokenUpdateTransactionToBytes tokenUpdateTransaction
+        let operations = Types.rawCborFromBytes bytes
+
+        tokenId <- case tokenIdFromText tokenIdText of
+            Right val -> return val
+            Left err -> logFatal ["Error couldn't parse token id:", err]
+
+        let payload = Types.TokenUpdate tokenId operations
+        let encodedPayload = Types.encodePayload payload
+
+        let nrgCost _ = return $ Just $ flip (tokenUpdateTransactionEnergyCost (Types.payloadSize encodedPayload) Cost.tokenAssignRevokeRolesCost) $ extendedCostFromOpts txOpts
+        txCfg <- liftIO $ getTransactionCfg baseCfg txOpts nrgCost
+
+        let intOpts = toInteractionOpts txOpts
+        let outFile = toOutFile txOpts
+        signAndProcessTransaction_ verbose txCfg encodedPayload intOpts outFile backend
+
+handlePLTUpdateMetadata ::
+    Backend ->
+    Maybe FilePath ->
+    Bool ->
+    Text ->
+    Maybe Text ->
+    Text ->
+    TransactionOpts (Maybe Types.Energy) ->
+    IO ()
+handlePLTUpdateMetadata backend baseCfgDir verbose metadataUrlText maybeChecksum tokenIdText txOpts = do
+    baseCfg <- getBaseConfig baseCfgDir verbose
+    when verbose $ do
+        runPrinter $ printBaseConfig baseCfg
+        putStrLn ""
+
+    metadata <- case maybeChecksum of
+        Just checksumStr ->
+            case parseChecksum checksumStr of
+                Nothing ->
+                    logFatal [printf "invalid checksum hash '%s'" checksumStr]
+                Just hash ->
+                    return $ CBOR.createTokenMetadataUrlWithSha256 metadataUrlText hash
+        Nothing ->
+            return $ CBOR.createTokenMetadataUrl metadataUrlText
+
+    withClient backend $ do
+        tokenOperation <- pure $ CBOR.TokenUpdateMetadata metadata
+        let tokenUpdateTransaction = CBOR.TokenUpdateTransaction (Seq.singleton tokenOperation)
+        let bytes = CBOR.tokenUpdateTransactionToBytes tokenUpdateTransaction
+        let operations = Types.rawCborFromBytes bytes
+
+        tokenId <- case tokenIdFromText tokenIdText of
+            Right val -> return val
+            Left err -> logFatal ["Error couldn't parse token id:", err]
+
+        let payload = Types.TokenUpdate tokenId operations
+        let encodedPayload = Types.encodePayload payload
+
+        let nrgCost _ = return $ Just $ flip (tokenUpdateTransactionEnergyCost (Types.payloadSize encodedPayload) Cost.tokenUpdateTokenMetadataCost) $ extendedCostFromOpts txOpts
         txCfg <- liftIO $ getTransactionCfg baseCfg txOpts nrgCost
 
         let intOpts = toInteractionOpts txOpts
@@ -1103,13 +1198,13 @@ handlePLTPausation backend baseCfgDir verbose pauseAction tokenIdText txOpts = d
 
         let tokenUpdateTransaction = CBOR.TokenUpdateTransaction (Seq.singleton tokenOperation)
         let bytes = CBOR.tokenUpdateTransactionToBytes tokenUpdateTransaction
-        let tokenParameter = Types.TokenParameter $ BS.toShort bytes
+        let operations = Types.rawCborFromBytes bytes
 
         tokenId <- case tokenIdFromText tokenIdText of
             Right val -> return val
             Left err -> logFatal ["Error couldn't parse token id:", err]
 
-        let payload = Types.TokenUpdate tokenId tokenParameter
+        let payload = Types.TokenUpdate tokenId operations
         let encodedPayload = Types.encodePayload payload
 
         let nrgCost _ = return $ Just $ flip (tokenUpdateTransactionEnergyCost (Types.payloadSize encodedPayload) Cost.tokenPauseUnpauseCost) $ extendedCostFromOpts txOpts
@@ -3173,9 +3268,7 @@ processConsensusCmd action _baseCfgDir verbose backend =
                             RequestFailed err -> logFatal ["Error getting token ID: I/O error: " <> err]
                     let maybeInitParams =
                             CBOR.tokenInitializationParametersFromBytes $
-                                BSBuilder.toLazyByteString $
-                                    BSBuilder.shortByteString $
-                                        Types.parameterBytes _cpltInitializationParameters
+                                Types.rawCborToLazyBytes _cpltInitializationParameters
                     case maybeInitParams of
                         Left _ -> do
                             logWarn ["Could not parse token initialization parameter bytes."]